Fix ns instantiate, but still not good.
